local/requirements.txt
pyserial
fire